network: preserve backwards compatibility
The new network parser will support specifying multiple networks in the old
format where each new non-indexed "lxc.network.type" line starts a new network
configuration. This way we don't break users. For now, we just print a
deprecation warning. We will KILL this in LXC 3.0.

+++ src/lxc/conf.h
@@ -124,7 +124,7 @@ union netdev_p {
  * @idx        : network counter
  */
 struct lxc_netdev {
-	unsigned int idx;
+	ssize_t idx;
 	int type;
 	int flags;
 	int ifindex;
